requirements.txt
torch
opencv-python
numpy
pandas
open-clip-torch
